cmakify build to prepare for using hunter (package manager)
Build with cmake instead of hard-coded bash-file, to be able to use
package manager.

Use toolchain file to customize build variables as is proper.

Setting up clang to use libc++ with cmake + toolchain also required
bumping ubuntu base image as cmake 3.10 was not available in the
existing one, and some more symlinking relate to clang was necessary to
get clang++-libc++ became necessary.